Code refactoring
Instead of having separate structures for each transport mode which have
the same data field refactor everything into single CoapEndpoint structure.
Remove unused methods.
