--- libcoap/src/transport/tls.rs
@@ -1,13 +1,0 @@
-// SPDX-License-Identifier: BSD-2-Clause
-/*
- * transport/dtls.rs - transport-specific code for TLS.
- * This file is part of the libcoap-rs crate, see the README and LICENSE files for
- * more information and terms of use.
- * Copyright © 2021-2023 The NAMIB Project Developers, all rights reserved.
- * See the README as well as the LICENSE file for more information.
- */
-
-/// TODO
-#[allow(dead_code)]
-#[cfg(feature = "tcp")]
-pub struct CoapTlsEndpoint {}
